+++ src/Situation.java
@@ -1,10 +1,10 @@
 public class Situation {
 
-    int X;
-    int Y;
-    int DIR;
+    float X;
+    float Y;
+    float DIR;
 
-    public Situation(int x, int y, int dir) {
+    public Situation(float x, float y, float dir) {
         X = x;
         Y = y;
         DIR = dir;
